backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
alembic>=1.13.0
passlib[bcrypt]>=1.7.4
bcrypt<5.0  # Pin to 4.x for compatibility with passlib 1.7.4
python-jose[cryptography]>=3.3.0
httpx>=0.28.0  # For testing